bot/requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
PyJWT==2.8.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
requests==2.31.0
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
